% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonogenic_assay_imaging_data.R
\name{clonogenic_assay_images}
\alias{clonogenic_assay_images}
\title{clonogenic_assay_images}
\usage{
clonogenic_assay_images()
}
\description{
Loads example images of colony growth in a clonogenic assay.
}
\examples{
clonogenic_assay_images()
}
