% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv_motility_imaging_data.R
\name{hiv_motility_images}
\alias{hiv_motility_images}
\title{hiv_motility_images}
\usage{
hiv_motility_images()
}
\description{
Loads example images of CD4+ T cells in a complex collagen environment.
}
\examples{
hiv_motility_images()
}
