% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular_particle_dynamics_data.R
\docType{data}
\name{subcellular_particle_dynamics}
\alias{subcellular_particle_dynamics}
\title{Subcellular particle dynamics data of labeled genomic loci in nuclei of genome engineered cell lines}
\usage{
data(subcellular_particle_dynamics)
}
\description{
This dataframe contains tracking data based on the imaging data: \code{\link{subcellular_particle_dynamcis_images}}.
Spots for the 15 first frames of the original dataset were identified and tracked with the FiJi plugin TrackMate.
}
\examples{
data(subcellular_particle_dynamics)
subcellular_particle_dynamics
}
\keyword{datasets}
