% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visustat_check.R
\name{visustat_check}
\alias{visustat_check}
\title{visustat_check}
\usage{
visustat_check(df, images, stack_time = FALSE, dimensions = 2)
}
\arguments{
\item{df}{dataframe of the form: \code{df(track, time, X, Y, (Z,) mapping_parameters, ...)}}

\item{images}{\code{vector}: filenames of images ordered by frames/time}

\item{dimensions}{\code{numeric}: specify whether the images are 2D or 3D.
If 3D is selected data is assumed to be in the form: \code{df(track, time, X, Y, Z, mapping paramters, ...)}}

\item{stack}{\code{logical}: default: \code{FALSE}, single image file provided if time-resolved imagestack is used, set: \code{TRUE}}
}
\description{
Gives a summary that allows to check the linking between a passed dataframe
 and images in a frame-wise manner.
}
\examples{
data(hiv_motility)
images <- hiv_motility_images()
visustat_check(hiv_motility, images)
}
