% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visustat_frame.R
\name{visustat_frame}
\alias{visustat_frame}
\title{visustat_frame}
\usage{
visustat_frame(df, ...)
}
\arguments{
\item{df}{dataframe of the form: \code{df(track, time, X, Y, (Z,) mapping_parameters, ...)}}

\item{image}{\code{character}: filename of image}

\item{stack}{\code{logical}: default: \code{FALSE}, single image file provided if time-resolved imagestack is used, set: \code{TRUE}}

\item{image.depth}{\code{numeric}: set image bit-depth; just important if Z-projections are calculated}

\item{image.normalize}{\code{logical}: normalize image}

\item{frame}{\code{integer}: frame to be mapped}

\item{tracks}{\code{vector}: defining tracks to be displayed}

\item{par.map}{\code{character}: specifying parameter in \code{df} to be visualized by color}

\item{par.shape}{\code{character}: specifying parameter in \code{df} to be mapped on shape}

\item{par.display}{display option for mapping; default: \code{TRUE}, mapping is disable with: \code{FALSE}}

\item{par.max}{\code{numeric}: defining upper range of color mapping}

\item{par.min}{\code{numeric}: defining lower range of color mapping}

\item{par.unit}{\code{character}: unit of the numeric mapped parameter}

\item{crop}{\code{logical}: option for cropping images; default: \code{FALSE}}

\item{sub.img}{\code{logical}: option for creating sub-images from specified \code{tracks} or pre-filtered \code{df}; default: \code{FALSE}}

\item{sub.window}{\code{numeric}: size of the sub-images in pixels}

\item{sub.col}{\code{numeric}: number of columns in which sub-images are arranged}

\item{tracks.size}{\code{numeric}: size of tracks}

\item{tracks.alpha}{\code{numeric}: transparency of tracks}

\item{tracks.length}{\code{numeric}: length of tracks (in frames)}

\item{tracks.label}{\code{logical}: when sub.img is used, display or hide track label}

\item{tracks.label.x}{\code{numeric}: when sub.img is used, set x-position of label}

\item{tracks.label.y}{\code{numeric}: when sub.img is used, set y-position of label}

\item{points.size}{\code{numeric}: size of points}

\item{points.alpha}{\code{numeric}: transparency of points}

\item{points.stat}{\code{character}: display statistic; default: \code{'echo'}, for blurring; without blurring \code{'identity'}}

\item{points.shape}{\code{numeric}: set shape from ggplot2 shape palette}

\item{axis.tick}{\code{numeric}: axis ticks in px}

\item{axis.display}{\code{logical}: display axis}

\item{axis.labs}{\code{logical}: display labs}

\item{unit}{\code{character}: setting name of unit; default: \code{'px'}}

\item{scaling}{\code{numeric}: scaling factor for unit; default: \code{1}}

\item{dimensions}{\code{numeric}: specify whether the images are 2D or 3D.
If 3D is selected the data is assumed to be in the form: \code{df(track, time, X, Y, Z, mapping paramters, ...)}}

\item{manual.z}{\code{numerice}: specify Z-plane to be visualized if no projection or sub windows are used}

\item{scale.bar}{\code{logical}: show scalebar; default: \code{FALSE}}

\item{scale.width}{\code{numeric}: width of scalebar; default: \code{40}}

\item{scale.height}{\code{numeric}: height of scalebar; default: \code{10}}

\item{scale.x}{\code{numeric}: distance from left border of the image towards scalebar}

\item{scale.y}{\code{numeric}: distance from bottom border of the image towards scalebar}

\item{scale.color}{\code{character}: specify color from R-color palette or hexcode}

\item{interactive}{\code{logical}: return the plot as an interactive plotly object. Not supported when using
sub.img or crop modes.}
}
\value{
returns a ggplot2 plot-object which can be further modified
}
\description{
With \code{visustat_frame}, continuous and discrete parameters can be mapped individually on color, shape and size for one timepoint.
}
\examples{
# import tracking data
data('hiv_motility')
# get image files
images <- hiv_motility_images()
# run visustat_frame with default settings
visustat_frame(hiv_motility, image=images[15], frame=15, image.normalize=1)
# run visustat_frame with specified settings
visustat_frame(hiv_motility,
   image = images[15],
   frame = 15,
   tracks = c(48, 66, 102, 108),
   sub.img = TRUE,
   sub.col = 2,
   sub.window= 300,
   par.map ='speed',
   par.shape ='type',
   points.size=2,
   image.normalize=1
 )
}
