% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visustat_shiny.R
\name{visustat_shiny}
\alias{visustat_shiny}
\title{visustat_shiny}
\usage{
visustat_shiny()
}
\description{
With visustat_shiny, you can run all functionalities of the VisuStatR package in a shiny app.
}
\examples{
visustat_shiny()
}
