% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visustat_summary.R
\name{visustat_summary}
\alias{visustat_summary}
\title{visustat_summary}
\usage{
visustat_summary(df, ...)
}
\arguments{
\item{df}{dataframe of the form: \code{df(track, time, X, Y, mapping_parameters, ...)}}

\item{frame}{\code{integer}: frame to be mapped}

\item{time.unit}{\code{character}: time unit}

\item{tracks}{\code{vector}: defining tracks to be displayed}

\item{par.map}{\code{character}: specifying parameter in \code{df} to be visualized by color}

\item{par.shape}{\code{character}: specifying parameter in \code{df} to be mapped on shape}

\item{ribbon}{\code{logical}: display ribbon}

\item{ribbon.stat}{\code{character}: choose ribbon-statistic: \code{'sd'}: standard deviation,
\code{'se'}: standard error,
\code{'ci'}: confidence interval}

\item{line.size}{\code{numeric}: size of lines}

\item{line.alpha}{\code{numeric}: transparency of lines}

\item{line.length}{\code{numeric}: length of lines (in frames)}

\item{points.size}{\code{numeric}: size of points}

\item{points.alpha}{\code{numeric}: transparency of points}

\item{points.shape}{\code{numeric}: set shape from ggplot2 shape palette}

\item{unit}{\code{character}: setting name of unit; default: \code{'px'}}
}
\description{
With \code{visustat_summary}, time-resolved summary statistics are calculated. In correspondence to \code{visustat_frame} continuous and discrete parameters can be mapped individually on color and shape.
 In addition, the population mean and either standard deviations, confidence intervals or standard errors for respective statistics can be displayed as facetted ribbon plots.
}
\examples{
data(hiv_motility)
visustat_summary(hiv_motility \%>\% na.omit(), par.map="speed", tracks=c(48, 66, 102, 108))
}
