% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonogenic_assay_data.R
\docType{data}
\name{clonogenic_assay}
\alias{clonogenic_assay}
\title{Clonogenic assay colony growth data}
\usage{
data(clonogenic_assay)
}
\description{
Description of dataset
}
\examples{
data(clonogenic_assay)
clonogenic_assay
}
\keyword{datasets}
