% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv_motility_data.R
\docType{data}
\name{hiv_motility}
\alias{hiv_motility}
\title{HIV motility data of HIV infected and uninfected CD4+ T cells in a loose collagen environment}
\usage{
data(hiv_motility)
}
\description{
Description of dataset
}
\examples{
data(hiv_motility)
}
\keyword{datasets}
