% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular_particle_dynamics_data.R
\docType{data}
\name{subcellular_particle_dynamics}
\alias{subcellular_particle_dynamics}
\title{Subcellular particle dynamics data}
\usage{
data(subcellular_particle_dynamics)
}
\description{
Description of dataset
}
\examples{
data(subcellular_particle_dynamics)
subcellular_particle_dynamics
}
\keyword{datasets}
