% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcellular_particle_dynamics_imaging_data.R
\name{subcellular_particle_dynamics_images}
\alias{subcellular_particle_dynamics_images}
\title{Subcellular particle dynamics of labeled genomic loci in engineered cell line}
\usage{
subcellular_particle_dynamics_images()
}
\description{
Loads example images of subcellular particle dynamics of labeled genomic loci in engineered cell line.
}
\examples{
subcellular_particle_dynamics_images()
}
